backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
Pillow
pytesseract
langchain
langchain-community
langchain-google-genai 
chromadb
numpy
python-dotenv
pdf2image
python-docx
gunicorn 